Conflict resolutions, moved generalization logic in electrum.rs
