+++ tests/integration/taproot_taker_contract_validation.rs
@@ -37,7 +37,7 @@ fn test_taproot_rejects_underfunded_maker_contract() {
     let maker_behaviors = vec![MakerBehavior::UnderfundTaprootContract];
 
     let (test_framework, mut takers, makers, block_generation_handle) =
-        TestFramework::init(makers_config_map, taker_behavior, maker_behaviors);
+        TestFramework::init::<BitcoindBackend>(makers_config_map, taker_behavior, maker_behaviors);
 
     let bitcoind = &test_framework.bitcoind;
     let taker = takers.get_mut(0).unwrap();
